% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_signal_val_data.R
\name{add_signal_val_data}
\alias{add_signal_val_data}
\title{add_signal_val_data}
\usage{
add_signal_val_data(
  data,
  group_1,
  group_2,
  treat_samps,
  lfc_samp_grna,
  grna_column
)
}
\arguments{
\item{data}{A data frame of normalized counts (at the gRNA level) for a set of samples with the true signal controlled.}

\item{group_1}{A character vector naming the columns that belong to group 1.}

\item{group_2}{A character vector naming the columns that belong to group 2.}

\item{treat_samps}{A character vector naming the treatment samples in \code{data}.}

\item{lfc_samp_grna}{A named vector of logFC samples, with gRNA ID names.}

\item{grna_column}{A character string naming the column containing gRNA IDs.}
}
\value{
A list containing the following components:
\itemize{
\item \code{data.bthin}: A data frame containing ground truth signal.
\item \code{coef_mat}: A coefficient matrix to be used by \code{seqgendiff}.
\item \code{design_mat}: A design matrix to be used by \code{seqgendiff}.
\item \code{thout}: output from \code{seqgendiff}.
\item \code{group_1}: A character vector of sample names in group_1.
\item \code{group_2}: A character vector of sample names in group_2.
\item 
}
}
\description{
For performance estimation, this function adds logFC signal to a known set of genes and guides in signal-controlled count data.
}
