% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_nonnull_lfc_estimates.R
\name{adjust_nonnull_lfc_estimates}
\alias{adjust_nonnull_lfc_estimates}
\title{adjust_nonnull_lfc_estimates}
\usage{
adjust_nonnull_lfc_estimates(res, filter_prop = 0.05)
}
\arguments{
\item{res}{An object produced by \code{delboy::get_crispr_gene_level_hits}.}

\item{filter_prop}{A real number (default = 0.05) specifying what proportion of the lowest-abundance genes to filter out prior to making non-null estimation.}
}
\value{
An object of class delboy_logfc
}
\description{
Applies adjustments to estimated non-null logFC estimates to ensure alignment with empirical observations.
}
