% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_combinations_treat_samples.R
\name{all_combinations_treat_samples}
\alias{all_combinations_treat_samples}
\title{all_combinations_treat_samples}
\usage{
all_combinations_treat_samples(samples, num_treat)
}
\arguments{
\item{samples}{A character vector of sample names.}

\item{num_treat}{An integer specifying the number of treatment samples.}
}
\value{
A matrix of combinations (each column corresponds to a separate combination). There will be \code{choose(length(samples), length(num_treat))} combinations (columns).
}
\description{
Returns all combinations of treatment samples for validation purposes.
}
