% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_fdr_thr_val_hits.R
\name{apply_fdr_thr_val_hits}
\alias{apply_fdr_thr_val_hits}
\title{apply_fdr_thr_val_hits}
\usage{
apply_fdr_thr_val_hits(data, filter_column, filter_thr, input_type = "hits")
}
\arguments{
\item{data}{A data frame as produced by \code{delboy::make_delboy_crispr_hit_comparison_table}.}

\item{filter_column}{A character string naming a column to be filtered.}

\item{filter_thr}{A numeric threshold for \code{filter_column}.}

\item{input_type}{A character string naming the input type: \code{hits} (default), or \code{all}.}
}
\value{
A data frame.
}
\description{
Applies an FDR threshold (for a given metric) to control the FDR for validation hits.
}
