% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_thresholds_val.R
\name{apply_thresholds_val}
\alias{apply_thresholds_val}
\title{apply_thresholds_val}
\usage{
apply_thresholds_val(data, thresh)
}
\arguments{
\item{data}{A data frame containing DEG results for multiple validation samples.}

\item{thresh}{An object produced by \code{delboy::find_pval_target_fdr}.}
}
\value{
A modified data frame.
}
\description{
Applies p-value and abs(LFC) thresholds to validation data.
}
