% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_elnet_hits.R
\name{assemble_elnet_hits}
\alias{assemble_elnet_hits}
\title{assemble_elnet_hits}
\usage{
assemble_elnet_hits(data, deseq2_res, decision_boundary)
}
\arguments{
\item{data}{Output from \code{delboy::combine_validation_original_hits}.}

\item{deseq2_res}{Output from \code{delboy::run_deseq2}.}

\item{decision_boundary}{A data frame providing co-ordinates for the SVM-learned decision boundary for FPs.}
}
\value{
A data frame.
}
\description{
Assembles Elastic-net hits together with log2 Fold Changes (from \code{DESeq2}) and an indication of whether a hit is predicted to be a false positive using an SVM fit.
}
