% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correct.R
\name{batch_correct}
\alias{batch_correct}
\title{batch_correct}
\usage{
batch_correct(data, batches, gene_column, method, parametric = FALSE)
}
\arguments{
\item{data}{A data frame of normalized count data.}

\item{batches}{A character vector naming the batches that the samples belong to.}

\item{gene_column}{A character string naming the column containing gene names.}

\item{method}{A character string naming the batch correction method: \code{combat_seq} or \code{combat}.}

\item{parametric}{Logical, whether to perform a parametric correction or not (defaults to \code{FALSE}).}
}
\value{
A data frame of batch-corrected data (can contain negative values).
}
\description{
Performs a batch correction on normalized count data using \code{sva}'s \code{ComBat} method.
}
