% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostX.R
\name{boostX}
\alias{boostX}
\title{boostX}
\usage{
boostX(data, diff_exp, group_1, group_2, gene_column, pvalue_column)
}
\arguments{
\item{data}{A data frame containing normalized count data for two different groups and their replicates. Can be a path to a file. Ideally the counts will already have been normalized to Transcripts Per Million (TPM), using, for example, the bias-aware quantification methods employed by \code{salmon} (Patro et al. 2017).}

\item{diff_exp}{Differential expression/representation results from a particular algorithm.}

\item{group_1}{A character string naming the columns that belong to group 1.}

\item{group_2}{A character string naming the columns that belong to group 2.}

\item{gene_column}{A character string naming the column containing gene names.}

\item{pvalue_column}{A character string naming the pvalue column in \code{diff_exp}.}

\item{max.iter}{An integer value indicating the maximum number of validation samples (default = 10). \code{NULL} indicates all sample combinations should be taken.}

\item{bcorr_data_validation}{\code{NULL} if no batch (signal) corrected data is already available for validation. Otherwise, a data frame of treatment-corrected data should be supplied (to speed up validation, if already available). Defaults to \code{NULL}. Batch correction will be conducted using \code{sva::ComBat} using non-parametric priors.}
}
\value{
A list containing the following elements:
}
\description{
Takes the output from a given differential expression/representation algorithm and tries to boost sensitivity while controlling the FDR.
}
\author{
Alex T. Kalinka
}
