% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fdr_threshold.R
\name{calc_fdr_threshold}
\alias{calc_fdr_threshold}
\title{calc_fdr_threshold}
\usage{
calc_fdr_threshold(data, metric_column, hit_class_column, target_fdr)
}
\arguments{
\item{data}{A data frame generated by \code{delboy::combine_harmonic_mean_pvals}.}

\item{metric_column}{A character string naming the column in \code{data} that contains a metric for which we wish to finf a threshold value.}

\item{hit_class_column}{A character string naming the column in \code{data} that contains \code{True_Positive} and \code{False_Positive} classifications.}

\item{target_fdr}{Numeric value (0-1) giving the target FDR. Defaults to 0.1.}
}
\value{
A data frame.
}
\description{
Calculates a threshold value for a given metric to achieve a specified FDR.
}
