% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_library_depth_median_ratio.R
\name{calc_median_normalization_size_factors}
\alias{calc_median_normalization_size_factors}
\title{calc_median_normalization_size_factors}
\usage{
calc_median_normalization_size_factors(counts)
}
\arguments{
\item{counts}{A data frame of counts for each sample in the study (samples as columns, gRNAs as rows).}
}
\value{
A data frame with 1 row and as many columns as \code{counts} with size factors for each experiment.
}
\description{
Calculates the size factors used in median normalization for count data: median of the geometric mean of counts across all experiments.
}
\author{
Alex T. Kalinka, \email{alex.kalinka@cancer.org.uk}
}
