% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_perf_pval_windows.R
\name{calc_perf_pval_windows}
\alias{calc_perf_pval_windows}
\title{calc_perf_pval_windows}
\usage{
calc_perf_pval_windows(
  data,
  pval_column,
  gene_column,
  abs_lfc_column,
  deg_genes
)
}
\arguments{
\item{data}{A data frame of DEG p-values.}

\item{pval_column}{A character string naming the p-value column.}

\item{gene_column}{A character string naming the gene column.}

\item{abs_lfc_column}{A character string naming the abs log fold change column.}

\item{deg_genes}{A character vector naming DEGs.}
}
\value{
A data frame of performance estimates across the p-val thresholds with log fold change values maximising the KS distance at each p-value (\code{abs_Log2FoldChange.argmax_KS_dist}).
}
\description{
Calculates sensitivity and FDR (after excluding predicted false positives) across a progressively larger window of DEG p-values.
}
