% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_perf_stats.R
\name{calc_perf_stats}
\alias{calc_perf_stats}
\title{calc_perf_stats}
\usage{
calc_perf_stats(data)
}
\arguments{
\item{data}{A data frame as produced by \code{delboy::make_delboy_crispr_hit_comparison_table}.}
}
\value{
A data frame.
}
\description{
Returns sensitivity and FDR estimates.
}
