% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_PRROC.R
\name{calc_PRROC}
\alias{calc_PRROC}
\title{calc_PRROC}
\usage{
calc_PRROC(data, score_col, group_col = NULL)
}
\arguments{
\item{data}{A data frame containing a 'TP' logical column indicating which genes are true positives.}

\item{score_col}{A character string naming a column containing a score metric.}

\item{group_col}{A character string naming the grouping column of 'data'. If \code{NULL} then no group, defaults to \code{NULL}.}
}
\value{
A data frame containing the following columns: \code{Precision}, \code{Recall}, \code{Sensitivity_FDR_10pct}, \code{Sensitivity_FDR_10pct}, and \code{AUPrRc}.
}
\description{
Returns AUPrRc, precision, and recall values for a data frame containing TP and a score metric.
}
