% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_perf_stats.R
\name{calculate_perf_stats}
\alias{calculate_perf_stats}
\title{calculate_perf_stats}
\usage{
calculate_perf_stats(perf_stats)
}
\arguments{
\item{perf_stats}{A data frame of performance stats for individual validation samples.}
}
\value{
A data frame.
}
\description{
Return sensitivity and precision estimates for delboy validation data as produced by \code{delboy::perf_stats_deg}.
}
