% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_harmonic_mean_pvals.R
\name{combine_harmonic_mean_pvals}
\alias{combine_harmonic_mean_pvals}
\title{combine_harmonic_mean_pvals}
\usage{
combine_harmonic_mean_pvals(data, pval_col, group_col, target_fdr)
}
\arguments{
\item{data}{A data frame containing p-values and a grouping variable.}

\item{pval_col}{A character string naming the p-value column.}

\item{group_col}{A character string naming the grouping column.}

\item{target_fdr}{A real number (0-1) specifying the FDR.}
}
\value{
An object of class delboy_logfc
}
\description{
Combines p-values for groups of dependent tests using a harmonic mean approach.
}
\references{
Wilson, D. J. 2019.The harmonic mean p-value for combining dependent tests. PNAS 116 (4): 1195-1200.
}
