% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_hits_orig_val_data.R
\name{combine_hits_orig_val_data}
\alias{combine_hits_orig_val_data}
\title{combine_hits_orig_val_data}
\usage{
combine_hits_orig_val_data(orig_hits, val_hits)
}
\arguments{
\item{orig_hits}{A data frame as produced by \code{delboy::get_crispr_gene_level_hits}.}

\item{val_hits}{A data frame as produced by \code{delboy::make_delboy_crispr_hit_comparison_table}.}
}
\value{
A data frame.
}
\description{
Combines hits from original and validation data for comparison purposes.
}
