% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_pos_neg_res.R
\name{combine_pos_neg_res}
\alias{combine_pos_neg_res}
\title{combine_pos_neg_res}
\usage{
combine_pos_neg_res(data_pos, data_neg)
}
\arguments{
\item{data_pos}{A data frame of positive hits as produced by \code{delboy::get_crispr_gene_level_hits}.}

\item{data_neg}{A data frame of negative hits as produced by \code{delboy::get_crispr_gene_level_hits}.}
}
\value{
A data frame.
}
\description{
Combines positive and negative gene-level CRISPR results.
}
