% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_validation_original_hits.R
\name{combine_validation_original_hits}
\alias{combine_validation_original_hits}
\title{combine_validation_original_hits}
\usage{
combine_validation_original_hits(
  elnet_lr_res,
  deseq2_res,
  delboy_validation_hits
)
}
\arguments{
\item{elnet_lr_res}{An object of class \code{delboy_elnet}, the output from running \code{delboy::run_elnet_logistic_reg}.}

\item{deseq2_res}{The output from \code{delboy::run_deseq2}.}

\item{delboy_validation_hits}{The output from running \code{delboy::make_delboy_hit_comparison_table}.}
}
\value{
A data frame.
}
\description{
Combine validation hits with hits from unmodified input data to aid analysis of false positives.
}
