% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{E2H}
\alias{E2H}
\title{E2H}
\description{
A sample of signal-corrected expression data from a bulk RNA experiment with signal added to a known set of genes.
}
\details{
A list with the following elements:
\itemize{
\item \code{data_bcorr}: A data frame of batch-corrected counts.
\item \code{data_val_bcorr}: A data frame of signal-corrected counts.
\item \code{lfc_samp}: A named vector of log-fold changes added to \code{data_bcorr} in which the names are gene names.
\item \code{group_1}: The samples that belong to group 1.
\item \code{group_2}: The samples that belong to group 2.
}
}
\keyword{data}
