% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_nonnull_logfc_distr.R
\name{estimate_nonnull_logfc_distr}
\alias{estimate_nonnull_logfc_distr}
\title{estimate_nonnull_logfc_distr}
\usage{
estimate_nonnull_logfc_distr(logfc, signed = FALSE)
}
\arguments{
\item{logfc}{A vector of logFC values derived from \code{delboy::run_deseq2}.}

\item{signed}{Logical indicating whether signed logFC distributions should be returned (defaults to \code{FALSE}).}
}
\value{
An object of class delboy_logfc
}
\description{
Estimates the full non-null logFC distribution based on \code{DESeq2} logFC estimates and mixture model fits employed by \code{locfdr}.
}
