% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_number_non_nulls.R
\name{estimate_number_non_nulls}
\alias{estimate_number_non_nulls}
\title{estimate_number_non_nulls}
\usage{
estimate_number_non_nulls(pvals)
}
\arguments{
\item{pvals}{A numeric vector of raw (unadjusted) p-values.}
}
\value{
A list containing an integer giving an estimate of the number of non-null cases, and an estimate of the misfit of the \code{locfdr} mixture model.
}
\description{
Estimates the number of non-null cases in a dataset using \code{locfdr} with raw p-values as inputs.
}
