% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_performance_boostx.R
\name{evaluate_performance_boostx}
\alias{evaluate_performance_boostx}
\title{evaluate_performance_boostx}
\usage{
evaluate_performance_boostx(
  data,
  group_1,
  group_2,
  gene_column,
  max.iter,
  num_non_null,
  lfc,
  lfc_dens,
  target_fdr
)
}
\arguments{
\item{data}{A data frame of normalized (and batch-corrected, if necessary) counts for a set of samples with the true signal controlled (by, for example, batch-correcting it).}

\item{group_1}{A character vector naming the columns that belong to group 1.}

\item{group_2}{A character vector naming the columns that belong to group 2.}

\item{gene_column}{A character string naming the column containing gene names.}

\item{max.iter}{An integer value indicating the maximum number of validation samples.}

\item{num_non_null}{An integer value indicating the number of genes to add signal to.}

\item{lfc}{A vector of logFC values for non-null cases.}

\item{lfc_dens}{A vector of density estimates for the logFC values given in \code{lfc}.}

\item{target_fdr}{A numerical value (0-1) indicating the target FDR.}
}
\value{
An object of class \code{boostx_performance}.
}
\description{
Evaluates the performance of \code{delboy} on RNAseq data by comparison with \code{DESeq2} output on the original input data for an experiment, controlling for real signal, and adding known signal (using \code{seqgendiff}'s binomial-thinning approach) for a sampled number of genes from a logFC distribution with both the number and distribution chosen to match as closely as possible the signal in the real data.
}
