% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_performance_crispr_calls.R
\name{evaluate_performance_crispr_calls}
\alias{evaluate_performance_crispr_calls}
\title{evaluate_performance_crispr_calls}
\usage{
evaluate_performance_crispr_calls(
  data,
  data_lfc,
  group_1,
  group_2,
  gene_column,
  grna_column,
  lfc_column,
  max.iter,
  num_non_null,
  lfc,
  lfc_dens,
  alt_hyp,
  target_fdr = 0.1
)
}
\arguments{
\item{data}{A data frame of normalized (and batch-corrected, if necessary) counts (at the gRNA level) for a set of samples with the true signal controlled (by, for example, batch-correcting it).}

\item{data_lfc}{A data frame of logFC estimates for \code{data}.}

\item{group_1}{A character vector naming the columns that belong to group 1.}

\item{group_2}{A character vector naming the columns that belong to group 2.}

\item{gene_column}{A character string naming the column containing gene names.}

\item{grna_column}{A character string naming the column containing gRNA IDs.}

\item{lfc_column}{A character string naming the logFC column in \code{data_lfc}.}

\item{max.iter}{An integer value indicating the maximum number of validation sample combinations..}

\item{num_non_null}{An integer value indicating the number of genes to add signal to.}

\item{lfc}{A vector of logFC values for non-null cases.}

\item{lfc_dens}{A vector of density estimates for the logFC values given in \code{lfc}.}

\item{alt_hyp}{A character string specifying the alternative hypothesis. Can be one of: \code{greater}, or \code{less} (see \code{DESeq2::results} documentation).}

\item{target_fdr}{Numeric value (0-1) giving the target FDR. Defaults to 0.1.}
}
\value{
An object of class \code{delboy_performance_crispr}.
}
\description{
Evaluates the performance of \code{delboy} on CRISPR pooled data by controlling for real signal, and adding known signal (using \code{seqgendiff}'s binomial-thinning approach) for a sampled number of genes from a logFC distribution with both the number and distribution chosen to match as closely as possible the signal in the real data.
}
