% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_predicted_FP_perf.R
\name{exclude_predicted_FP_perf}
\alias{exclude_predicted_FP_perf}
\title{exclude_predicted_FP_perf}
\usage{
exclude_predicted_FP_perf(data_svm, perf_stats, num_pos)
}
\arguments{
\item{data_svm}{The \code{data_svm} element from \code{delboy::svm_false_positive_classification}.}

\item{perf_stats}{The \code{performance_stats} from \code{delboy::evaluate_performance_rnaseq_calls}.}

\item{num_pos}{An integer giving the number of true positives in total.}
}
\value{
A data frame.
}
\description{
Performance stats with predicted false positives removed (can impact both sensitivity and FDR/precision).
}
