% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_logfc_guides.R
\name{expand_logfc_guides}
\alias{expand_logfc_guides}
\title{expand_logfc_guides}
\usage{
expand_logfc_guides(data, gene_col, lfc_col, med_lfc, lfc_half_window = 0.3)
}
\arguments{
\item{data}{A data frame containing gene and logFC columns at the gRNA level.}

\item{gene_col}{A character string naming the gene column in \code{data}.}

\item{lfc_col}{A character string naming the logFC column in \code{data}.}

\item{med_lfc}{A vector of logFC values (originally sampled from an estimated non-null logFC distribution derived from \code{data}).}

\item{lfc_half_window}{A real number providing the size of the half-window around each logFC when defining the sampling population. Defaults to 0.3.}
}
\value{
A data frame containing a gene and logFC column with sampled logFC values.
}
\description{
Expands a set of summary per-gene logFC values into within-gene per-guide level logFC values by sampling from original guide-level logFC data.
}
