% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_grid_decision_boundary.R
\name{extract_grid_decision_boundary}
\alias{extract_grid_decision_boundary}
\title{extract_grid_decision_boundary}
\usage{
extract_grid_decision_boundary(grid)
}
\arguments{
\item{grid}{A data frame with the following columns: \code{abs_log2FoldChange}, \code{log10_baseExpr}, and \code{Predicted_FP}.}
}
\value{
A data frame.
}
\description{
Reduces a grid of class predictions to a decision boundary.
}
