% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_pval_target_fdr.R
\name{find_pval_target_fdr}
\alias{find_pval_target_fdr}
\title{find_pval_target_fdr}
\usage{
find_pval_target_fdr(data, target_FDR)
}
\arguments{
\item{data}{A data frame containing FDR estimates and the corresponding unadjusted p-values.}

\item{target_FDR}{A numerical value (0-1) giving the target FDR.}
}
\value{
A list containing the following elements:
\itemize{
\item \code{pvalue} A numerical value giving the p-value corresponding to the target FDR.
\item \code{loess_fit_fdr_all} An object of class \code{loess}.
\item \code{loess_fit_fdr_excl} An object of class \code{loess}.
\item \code{loess_fit_sens_all} An object of class \code{loess}.
\item \code{loess_fit_sens_excl} An object of class \code{loess}.
\item \code{target_FDR} A numerical value (0-1) giving the target FDR.
}
}
\description{
Returns the unadjusted p-value that corresponds to a target FDR using a LOESS fit to the FDR estimates.
}
