% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crispr_gene_level_hits.R
\name{get_crispr_gene_level_hits}
\alias{get_crispr_gene_level_hits}
\title{get_crispr_gene_level_hits}
\usage{
get_crispr_gene_level_hits(
  data,
  grna_column,
  gene_column,
  group_1,
  group_2,
  target_fdr = 0.1,
  dir = "both"
)
}
\arguments{
\item{data}{A data frame of normalized counts at the gRNA level.}

\item{grna_column}{A character string specifying the gRNA ID column.}

\item{gene_column}{A character string specifying the gene name column.}

\item{group_1}{A character vector naming the columns that belong to group 1.}

\item{group_2}{A character vector naming the columns that belong to group 2.}

\item{target_fdr}{Numeric value (0-1) giving the target FDR. Defaults to 0.1.}

\item{dir}{A character string specifying which end of the fold change distribution to test: \code{both} (default), \code{pos}, \code{neg}.}
}
\value{
An object of class \code{delboy_crispr_hits} which is a list containing the following components:
\itemize{
\item \code{deseq2_pos}: A data frame containing positive \code{DESeq2} results at the gRNA level.
\item \code{hmp_gene_pos}: A data frame containing positive gene-level results (with p-values derived from the harmonic mean p-value method).
\item \code{deseq2_neg}: A data frame containing negative \code{DESeq2} results at the gRNA level.
\item \code{hmp_gene_neg}: A data frame containing negative gene-level results (with p-values derived from the harmonic mean p-value method).
}
}
\description{
Runs \code{DESeq2} at the gRNA-level and uses a harmonic mean p-value method to combine the p-values to the gene-level - extracts hits at both the positive and negative ends.
}
\references{
Wilson, D. J. 2019.The harmonic mean p-value for combining dependent tests. PNAS 116 (4): 1195-1200.
}
