% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deseq2_results.R
\name{get_deseq2_results}
\alias{get_deseq2_results}
\title{get_deseq2_results}
\usage{
get_deseq2_results(delboy_res)
}
\arguments{
\item{delboy_res}{An object of class \code{delboy}.}
}
\value{
A data frame.
}
\description{
Retrieve a data frame of \code{DESeq2} results from an object of class \code{delboy}.
}
