% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_performance_stats.R
\name{get_performance_stats}
\alias{get_performance_stats}
\title{get_performance_stats}
\usage{
get_performance_stats(delboy_res)
}
\arguments{
\item{delboy_res}{An object of class \code{delboy}.}
}
\value{
A data frame.
}
\description{
Extract performance stats from an object of class \code{delboy}.
}
