% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_metric_for_thr.R
\name{infer_metric_for_thr}
\alias{infer_metric_for_thr}
\title{infer_metric_for_thr}
\usage{
infer_metric_for_thr(data)
}
\arguments{
\item{data}{A data frame generated by \code{delboy::make_delboy_crispr_hit_comparison_table}.}
}
\value{
A list containing the following components:
\itemize{
\item \code{summ_stats}: A data frame summarising median values of \code{mean_log2FoldChange} and \code{sd_log2FoldChange} for TPs and FPs.
\item \code{metric}: A character string, either: \code{mean_log2FoldChange}, \code{sd_log2FoldChange}, or \code{NA} (no metric to be used for FDR thresholding).
}
}
\description{
For performance estimation, this function infers whether \code{mean_log2FoldChange} or \code{sd_log2FoldChange} should be used for any FDR thresholding.
}
