% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfc_nonnull_kdensity.R
\name{lfc_nonnull_kdensity}
\alias{lfc_nonnull_kdensity}
\title{lfc_nonnull_kdensity}
\usage{
lfc_nonnull_kdensity(logfc, lfdr)
}
\arguments{
\item{logfc}{A vector of logFC values.}

\item{lfdr}{The return object from \code{locfdr::locfdr}.}
}
\value{
A list.
}
\description{
Returns a kernel density estimate for the logFC of non-null cases.
}
