% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_coef_matrix.R
\name{make_coef_matrix}
\alias{make_coef_matrix}
\title{make_coef_matrix}
\usage{
make_coef_matrix(data, lfc_samp, gene_column)
}
\arguments{
\item{data}{A data frame of normalized (and batch-corrected, if necessary) counts for a set of samples.}

\item{lfc_samp}{A named vector of logFC values where names are gene names.}

\item{gene_column}{A character string naming the column containing gene names.}
}
\value{
A coefficient matrix.
}
\description{
Make a coefficient matrix for use by \code{seqgendiff}.
}
