% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_delboy_crispr_hit_comparison_table.R
\name{make_delboy_crispr_hit_comparison_table}
\alias{make_delboy_crispr_hit_comparison_table}
\title{make_delboy_crispr_hit_comparison_table}
\usage{
make_delboy_crispr_hit_comparison_table(comb_pvals, lfc_samp)
}
\arguments{
\item{comb_pvals}{A data frame generated by \code{delboy::combine_harmonic_mean_pvals}.}

\item{lfc_samp}{A named vector of logFC values where the names are gene names. These are the true positives with their associated logFC values. All other genes are true negatives.}
}
\value{
A data frame.
}
\description{
Make a TP, FN, FP data frame to aid analysis of \code{delboy} validation hits produced by \code{delboy::evaluate_performance_crispr_calls}.
}
