% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_delboy_hit_comparison_table.R
\name{make_delboy_hit_comparison_table}
\alias{make_delboy_hit_comparison_table}
\title{make_delboy_hit_comparison_table}
\usage{
make_delboy_hit_comparison_table(elnet_lr_res, deseq2_res, lfc_samp)
}
\arguments{
\item{elnet_lr_res}{An object of class \code{delboy_elnet}, the output from running \code{delboy::run_elnet_logistic_reg}.}

\item{deseq2_res}{The output from \code{delboy::run_deseq2}.}

\item{lfc_samp}{A named vector of logFC values where the names are gene names. These are the true positives with their associated logFC values. All other genes are true negatives.}
}
\value{
A data frame.
}
\description{
Make a TP, FN, FP data frame to aid analysis of \code{delboy} validation hits produced by \code{delboy::evaluate_performance_rnaseq_calls}.
}
