% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ground_truth_data.R
\name{make_ground_truth_data}
\alias{make_ground_truth_data}
\title{make_ground_truth_data}
\usage{
make_ground_truth_data(
  data,
  ctrl,
  treat,
  grna_column,
  gene_column,
  num_non_null,
  treat_sig,
  data_lfc = NULL,
  data.sc = NULL,
  el = NULL,
  method = "combat_seq",
  normalize = FALSE
)
}
\arguments{
\item{data}{A data frame of counts}

\item{ctrl}{A character string naming the control sample columns.}

\item{treat}{A character string naming the treatment sample columns.}

\item{grna_column}{A character string specifying the gRNA ID column.}

\item{gene_column}{A character string naming the gene column in \code{data}.}

\item{num_non_null}{An integer specifying the number of non-null cases to add signal to.}

\item{treat_sig}{The samples to be treated as treatment samples.}

\item{data_lfc}{Output from \code{delboy::get_crispr_gene_level_hits}. If \code{NULL} (default) this will be produced within the function.}

\item{data.sc}{Output from \code{delboy::prep_val_data}. If \code{NULL} (default) this will be produced within the function.}

\item{el}{Output from \code{delboy::adjust_nonnull_lfc_estimates}. If \code{NULL} (default) this will be produced within the function.}

\item{method}{A character string naming the batch correction method: \code{combat_seq} (default) or \code{combat}.}

\item{normalize}{Logical, whether to normalize the data or not (default = \code{FALSE}).}
}
\value{
A list.
}
\description{
Add signal to specified num of non-null genes (evenly split across pos and neg cases).
}
