% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_treat_factor.R
\name{make_treat_factor}
\alias{make_treat_factor}
\title{make_treat_factor}
\usage{
make_treat_factor(data, group_1, group_2)
}
\arguments{
\item{data}{A matrix of counts with named columns.}

\item{group_1}{A character string naming the columns that belong to group 1.}

\item{group_2}{A character string naming the columns that belong to group 2.}
}
\value{
A factor vector of treatments corresponding to the columns of the input \code{data}.
}
\description{
Returns a treatment factor for use in batch-correction and differential-expression analyses.
}
