% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_up_FPs.R
\name{mark_up_FPs}
\alias{mark_up_FPs}
\title{mark_up_FPs}
\usage{
mark_up_FPs(res, filter_column, filter_thr)
}
\arguments{
\item{res}{Output from \code{delboy::get_crispr_gene_level_hits}.}

\item{filter_column}{A character string naming the metric against which FPs will be defined.}

\item{filter_thr}{A numeric threshold value to be applied to \code{filter_column}.}
}
\value{
A data frame.
}
\description{
Combines hits from original and validation data for comparison purposes.
}
