% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_diff_val_samples.R
\name{max_diff_val_samples}
\alias{max_diff_val_samples}
\title{max_diff_val_samples}
\usage{
max_diff_val_samples(all_comb, max.iter)
}
\arguments{
\item{all_comb}{A matrix, as produced by \code{delboy::all_combinations_treat_samples}.}

\item{max.iter}{An integer value indicating the maximum number of validation sample combinations.}
}
\value{
A matrix of validation 'treatment' sample combinations.
}
\description{
Extracts a maximally distinct set of treatment samples for performance validation.
}
