% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_library_depth_median_ratio.R
\name{normalize_library_depth_median_ratio}
\alias{normalize_library_depth_median_ratio}
\title{normalize_library_depth_median_ratio}
\usage{
normalize_library_depth_median_ratio(counts)
}
\arguments{
\item{counts}{A data frame of counts for each sample in the study (samples as columns, gRNAs as rows).}
}
\value{
A data frame in which the sample columns are library-depth normalized.
}
\description{
Normalizes library depth from a counts file using the median-ratio method \link{1}.
}
\references{
Wang T, Wei JJ, Sabatini DM, Lander ES. Genetic screens in human cells using the CRISPR-Cas9 system. Science. 2014, 343: 80-84.
Anders S, Huber W. Differential expression analysis for sequence count data. Genome Biol. 2010, 11: R106
}
\author{
Alex T. Kalinka, \email{alex.kalinka@cancer.org.uk}
}
