% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_library_depth_relative.R
\name{normalize_library_depth_relative}
\alias{normalize_library_depth_relative}
\title{normalize_library_depth_relative}
\usage{
normalize_library_depth_relative(counts, norm_total)
}
\arguments{
\item{counts}{A data frame of counts for each sample in the study (samples as columns, gRNAs as rows).}

\item{norm_total}{An integer specifying the fixed total number of reads each sample will be set to (e.g. 2e7). If \code{NULL}, then the median total across the samples is used.}
}
\value{
A data frame in which the sample columns are library-depth normalized.
}
\description{
Normalizes library depth from a counts file by dividing the counts of each sample by its total and multiplying all samples by a constant total number of reads.
}
