% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf_stats_deg.R
\name{perf_stats_deg}
\alias{perf_stats_deg}
\title{perf_stats_deg}
\usage{
perf_stats_deg(delboy_res, deseq2_res, lfc_samp, padj_cutoff = 0.1)
}
\arguments{
\item{delboy_res}{An object of class \code{delboy_elnet} as produced by \code{delboy::run_elnet_logistic_reg}.}

\item{deseq2_res}{The output from \code{delboy::run_deseq2}.}

\item{lfc_samp}{A named vector of logFC values where the names are gene names. These are the true positives with their associated logFC values. All other genes are true negatives.}

\item{padj_cutoff}{A numeric value between 0 and 1 indicating the adjusted p-value cutoff to use for \code{DESeq2} hits. Defaults to 0.1.}
}
\value{
A data frame of sensitivity and precision estimates.
}
\description{
Extracts performance stats for differential expression calls made by \code{delboy} and \code{DESeq2}.
}
