% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_FP_delboy.R
\name{predict_FP_delboy}
\alias{predict_FP_delboy}
\title{predict_FP_delboy}
\usage{
predict_FP_delboy(data, db)
}
\arguments{
\item{data}{A data frame with the following columns: \code{log10_baseExpr} and \code{abs_log2FoldChange}.}

\item{db}{A data frame providing a decision boundary with the following columns: \code{log10_baseExpr} and \code{abs_log2FoldChange}.}
}
\value{
A data frame.
}
\description{
Predict false positives in expression data.
}
