% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_bthin_matrix_diffrep.R
\name{prep_bthin_matrix_diffrep}
\alias{prep_bthin_matrix_diffrep}
\title{prep_bthin_matrix_diffrep}
\usage{
prep_bthin_matrix_diffrep(
  data,
  bthin_matrix,
  sample_names,
  treat_cols,
  gene_column
)
}
\arguments{
\item{data}{A data frame of normalized (and batch-corrected, if necessary) counts for a set of samples with the true signal controlled (by, for example, batch-correcting it).}

\item{bthin_matrix}{The matrix output from \code{seqgendiff::thin_diff}.}

\item{sample_names}{A character vector giving the sample names from the original input data provided to \code{seqgendiff::thin_diff} in their original order.}

\item{treat_cols}{A vector of 1/0 where a 1 indicates a treatment sample in \code{seqgendiff::thin_diff}.}

\item{gene_column}{A character string naming the column containing gene names.}
}
\value{
A data frame.
}
\description{
Prepare the matrix output from \code{seqgendiff::thin_diff} for use in differential representation analyses by, for example, \code{DESeq2} or \code{delboy}.
}
