% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_val_data.R
\name{prep_val_data}
\alias{prep_val_data}
\title{prep_val_data}
\usage{
prep_val_data(
  data,
  group_1,
  group_2,
  grna_column,
  gene_column,
  bc_method = "combat_seq",
  normalize_method = "relative"
)
}
\arguments{
\item{data}{A data frame of counts (at the gRNA level) for a set of samples.}

\item{group_1}{A character vector naming the columns that belong to group 1.}

\item{group_2}{A character vector naming the columns that belong to group 2.}

\item{grna_column}{A character string naming the column containing gene names.}

\item{gene_column}{A character string naming the column containing gene names.}

\item{bc_method}{A character string naming the batch correction method: \code{combat_seq} (default) or \code{combat}.}

\item{normalize_method}{A character string naming the read depth normalization method: \code{relative} (default) or \code{median_ratio}. \code{NULL} indicates no normalization.}
}
\value{
A data frame of normalized, signal-corrected data.
}
\description{
Normalizes and batch-corrects signal in count data in preparation for use in performance evaluation.
}
