% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.delboy.R
\name{print.delboy}
\alias{print.delboy}
\title{print.delboy}
\usage{
\method{print}{delboy}(delboy_res)
}
\arguments{
\item{delboy_res}{Output from \code{delboy::run_delboy}.}

\item{...}{Other arguments to be passed to \code{print}.}
}
\value{
Prints to the console.
}
\description{
Prints summary of a \code{delboy} object to the console.
}
