% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_decision_boundary.R
\name{process_decision_boundary}
\alias{process_decision_boundary}
\title{process_decision_boundary}
\usage{
process_decision_boundary(data.grid)
}
\arguments{
\item{data.grid}{A grid of predicted true and false positives along the plane formed by absolute log fold change and log10 expression.}
}
\value{
A data frame.
}
\description{
Extract and smooth a decision boundary along the axes of fold-change and expression such that fold change is a convex, monotonically decreasing function of expression level.
}
