% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_delboy.R
\name{run_delboy}
\alias{run_delboy}
\title{run_delboy}
\usage{
run_delboy(
  data,
  group_1,
  group_2,
  normalize,
  filter_cutoff,
  gene_column,
  batches = NULL,
  batch_corr_method = "combat_np",
  bcorr_data_validation = FALSE,
  crispr = FALSE,
  grna_column = "sgRNA"
)
}
\arguments{
\item{data}{A data frame containing count data for two different groups and their replicates. Can be a path to a file. If the count data needs to be normalized, indicate this using the \code{normalize} argument.}

\item{group_1}{A character string naming the columns that belong to group 1.}

\item{group_2}{A character string naming the columns that belong to group 2.}

\item{normalize}{Character string naming the count normalization method. Can be one of \code{median_ratio}, \code{relative}, or \code{NULL} (indicates the data has already been normalized). If the input is RNAseq data, ideally the counts will already have been normalized to Transcripts Per Million (TPM), using, for example, the bias-aware quantification methods employed by \code{salmon} (Patro et al. 2017).}

\item{filter_cutoff}{A numerical value indicating the cutoff below which (summed across all replicates) a gene (or gRNA) will be removed from the data. For example, to keep only genes with more than 1 TPM on average across both groups, set the cutoff to 10 if there are 10 replicates in total.}

\item{gene_column}{A character string naming the column containing gene names.}

\item{batches}{A character vector identifying the batch structure of the experiment. The length must equal \code{length(group_1) + length(group_2)}. If \code{NULL}, there are no batches. Defaults to \code{NULL}.}

\item{batch_corr_method}{A character string naming the batch correction method. Can be one of \code{combat_np} (non-parametric) or \code{combat_p} (parametric). Defaults to \code{combat_np}. Ignored if \code{batches} is \code{NULL}.}

\item{bcorr_data_validation}{\code{NULL} if no batch corrected data is already available (and a batch correction of the true signal is required for validation). Otherwise, a data frame of treatment-corrected data should be supplied (to speed up validation, if already available).}

\item{crispr}{Logical - is the data from a CRISPR pooled screen. Defaults to \code{FALSE}.}

\item{grna_column}{A character string naming the column containing sgRNA IDs. Defaults to \code{sgRNA}. Ignored if \code{crispr} set to \code{FALSE}.}
}
\value{
An object of class \code{delboy}.
}
\description{
Performs a differential-representation analysis using an elastic-net logistic regression approach for (normalized) count data that is split into two groups.
}
\references{
Kalinka, A. T. 2020.
Patro, R. et al. 2017. Salmon provides fast and bias-aware quantification of transcript expression. Nature Methods 14: 417-419.
}
