% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_delboy_crispr.R
\name{run_delboy_crispr}
\alias{run_delboy_crispr}
\title{run_delboy_crispr}
\usage{
run_delboy_crispr(
  data,
  controls,
  treatments,
  grna_column,
  gene_column,
  target_fdr = 0.1,
  normalize_method = "relative",
  estim_lfc_distr = FALSE
)
}
\arguments{
\item{data}{A data frame containing count data for two different groups and their replicates.}

\item{controls}{A character string naming the control sample columns.}

\item{treatments}{A character string naming the treatment sample columns.}

\item{grna_column}{A character string naming the column containing gene names.}

\item{gene_column}{A character string naming the column containing gene names.}

\item{target_fdr}{Numeric value (0-1) giving the target FDR. Defaults to 0.1.}

\item{normalize_method}{A character string naming the read depth normalization method: \code{relative} (default) or \code{median_ratio}.}

\item{estim_lfc_distr}{Logical indicating whether to estimate the non-null log-fold change distribution in the data (default = \code{FALSE}).}
}
\value{
An object of class \code{delboy_crispr}.
}
\description{
Calculates a threshold value for a given metric to achieve a specified FDR.
}
