% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_deseq2.R
\name{run_deseq2}
\alias{run_deseq2}
\title{run_deseq2}
\usage{
run_deseq2(data, group_1, group_2, gene_column)
}
\arguments{
\item{data}{A data frame containing normalized count data.}

\item{group_1}{A character vector naming the columns that belong to group 1.}

\item{group_2}{A character vector naming the columns that belong to group 2.}

\item{gene_column}{A character string naming the column containing gene names.}
}
\value{
A data frame containing \code{DESeq2} results.
}
\description{
Runs \code{DESeq2} on normalized, batch-corrected data containing two groups to be contrasted.
}
