% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_elnet_logistic_reg.R
\name{run_elnet_logistic_reg}
\alias{run_elnet_logistic_reg}
\title{run_elnet_logist_reg}
\usage{
run_elnet_logistic_reg(data, treat, alpha)
}
\arguments{
\item{data}{A matrix of normalized, batch-corrected counts - samples as rows, genes as columns.}

\item{treat}{A factor vector grouping samples into one of at most two groups.}

\item{alpha}{The elastic-net regression penalty, between 0 and 1.}
}
\value{
An object of class \code{delboy_elnet}.
}
\description{
Run an Elastic-net logistic regression using \code{glmnet}.
}
