% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_lfc_genes_guides.R
\name{sample_lfc_genes_guides}
\alias{sample_lfc_genes_guides}
\title{sample_lfc_genes_guides}
\usage{
sample_lfc_genes_guides(
  data_lfc,
  num_non_null,
  gene_column,
  lfc_column,
  lfc,
  lfc_dens
)
}
\arguments{
\item{data_lfc}{A data frame containing logFC estimates at the gRNA level.}

\item{num_non_null}{An integer value indicating the number of genes to add signal to.}

\item{gene_column}{A character string naming the column containing gene names.}

\item{lfc_column}{A character string naming the logFC column in \code{data_lfc}.}

\item{lfc}{A vector of logFC values for non-null cases.}

\item{lfc_dens}{A vector of density estimates for the logFC values given in \code{lfc}.}

\item{grna_column}{A character string naming the column containing gRNA IDs.}
}
\value{
A list containing the following components:
\itemize{
\item \code{lfc_samp}: A named vector of logFC values, where names are gene names.
\item \code{lfc_samp_grna}: A named vector of logFC values, where names are gRNA IDs.
\item \code{lfc_samp_df}: A data frame of sampled logFC values at the gRNA level.
}
}
\description{
For performance estimation, this function samples logFC values, as well as genes and guides to add this signal to.
}
