% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_decision_boundary.R
\name{smooth_decision_boundary}
\alias{smooth_decision_boundary}
\title{smooth_decision_boundary}
\usage{
smooth_decision_boundary(db, entry_point = 1.25)
}
\arguments{
\item{db}{A data frame providing the TP-FP decision boundary for \code{log10_baseExpr} and \code{abs_log2FoldChange} columns.}

\item{entry_point}{Internal point to start smoothing in terms of \code{log10_baseExpr}. Defaults to 1.25.}
}
\value{
A data frame of smoothed values
}
\description{
Smoothes the fold-change expression decision boundary such that it is a convex, monotonically decreasing function of expression level.
}
