% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_filter_props_nnull_lfc.R
\name{try_filter_props_nnull_lfc}
\alias{try_filter_props_nnull_lfc}
\title{try_filter_props_nnull_lfc}
\usage{
try_filter_props_nnull_lfc(res, filter_prop)
}
\arguments{
\item{res}{An object produced by \code{delboy::get_crispr_gene_level_hits}.}

\item{filter_prop}{A real number (default = 0.05) specifying what proportion of the lowest-abundance genes to filter out prior to making non-null estimation.}
}
\value{
An object of class delboy_logfc
}
\description{
Applies adjustments to estimated non-null logFC estimates to ensure alignment with empirical observations.
}
