% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_elnet_hits.R
\name{assemble_elnet_hits}
\alias{assemble_elnet_hits}
\title{assemble_elnet_hits}
\usage{
assemble_elnet_hits(data, deseq2_res, svm_fit, svm_method)
}
\arguments{
\item{data}{Output from \code{delboy::combine_validation_original_hits}.}

\item{deseq2_res}{Output from \code{delboy::run_deseq2}.}

\item{svm_fit}{Output from \code{delboy::svm_false_positive_classification}.}

\item{svm_method}{A character string naming the SVM method. Can be one of \code{e1071} or \code{svmpath}.}
}
\value{
A data frame.
}
\description{
Assembles Elastic-net hits together with log2 Fold Changes (from \code{DESeq2}) and an indication of whether a hit is predicted to be a false positive using an SVM fit.
}
