% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_nonnull_logfc_distr.R
\name{estimate_nonnull_logfc_distr}
\alias{estimate_nonnull_logfc_distr}
\title{estimate_nonnull_logfc_distr}
\usage{
estimate_nonnull_logfc_distr(logfc)
}
\arguments{
\item{logfc}{A vector of logFC values.}
}
\value{
A list with the following elements:
\itemize{
\item \code{non_null.dens}: non-null density estimates.
\item \code{non_null.lfc}: logFC mid-points corresponding to the densities.
\item \code{misfit}: A numerical estimate of the misfit (larger values indicate worse misfits).
}
}
\description{
Estimates the full non-null logFC distribution based on \code{DESeq2} logFC estimates and mixture model fits employed by \code{locfdr}.
}
