% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ground_truth_data.R
\name{make_ground_truth_data}
\alias{make_ground_truth_data}
\title{make_ground_truth_data}
\usage{
make_ground_truth_data(
  data,
  ctrl,
  treat,
  grna_column,
  gene_column,
  num_non_null,
  treat_sig
)
}
\arguments{
\item{data}{A data frame of counts}

\item{ctrl}{A character string naming the control sample columns.}

\item{treat}{A character string naming the treatment sample columns.}

\item{grna_column}{A character string specifying the gRNA ID column.}

\item{gene_column}{A character string naming the gene column in \code{data}.}

\item{num_non_null}{An integer specifying the number of non-null cases to add signal to.}

\item{treat_sig}{The samples to be treated as treatment samples.}
}
\value{
A list.
}
\description{
Add signal to specified num of non-null genes (evenly split across pos and neg cases).
}
