% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_up_FPs.R
\name{mark_up_FPs}
\alias{mark_up_FPs}
\title{mark_up_FPs}
\usage{
mark_up_FPs(res, dir, mlfc_fp_thresh = 0.3)
}
\arguments{
\item{res}{Output from \code{delboy::combine_harmonic_mean_pvals}.}

\item{dir}{A character string identifying which end of logFC values we are focusing on: \code{less} or \code{greater}.}

\item{mlfc_fp_thresh}{Numeric indicating the scaled log-fold change threshold below which a hit should be considered a False Positive (default = 0.3).}
}
\value{
A data frame.
}
\description{
Combines hits from original and validation data for comparison purposes.
}
