% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.delboy.R
\name{plot.delboy}
\alias{plot.delboy}
\title{plot.delboy}
\usage{
\method{plot}{delboy}(delboy_res, type, ...)
}
\arguments{
\item{delboy_res}{Output from \code{delboy::run_delboy}.}

\item{type}{A character string naming the plot type: \code{fc_expr}.}

\item{...}{Other arguments to be passed to \code{plot}.}
}
\value{
Used for side-effect of plotting.
}
\description{
Plotting for \code{delboy} objects.
}
