% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_elnet_data.R
\name{prep_elnet_data}
\alias{prep_elnet_data}
\title{prep_elnet_data}
\usage{
prep_elnet_data(data, group_1, group_2, gene_column)
}
\arguments{
\item{data}{A data frame produced by \code{delboy::prep_bthin_matrix_diffrep}.}

\item{group_1}{A character vector naming the columns that belong to group 1.}

\item{group_2}{A character vector naming the columns that belong to group 2.}

\item{gene_column}{A character string naming the column containing gene names.}
}
\value{
A data frame.
}
\description{
Prepare data for use in an Elastic-net logistic regression.
}
