% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_delboy.R
\name{run_delboy}
\alias{run_delboy}
\title{run_delboy}
\usage{
run_delboy(
  data,
  group_1,
  group_2,
  filter_cutoff,
  gene_column,
  batches = NULL,
  batch_corr_method = "combat_np",
  bcorr_data_validation = NULL
)
}
\arguments{
\item{data}{A data frame containing normalized count data for two different groups and their replicates. Can be a path to a file. Ideally the counts will already have been normalized to Transcripts Per Million (TPM), using, for example, the bias-aware quantification methods employed by \code{salmon} (Patro et al. 2017).}

\item{group_1}{A character string naming the columns that belong to group 1.}

\item{group_2}{A character string naming the columns that belong to group 2.}

\item{filter_cutoff}{A numerical value indicating the cutoff below which (summed across all replicates) a gene (or gRNA) will be removed from the data. For example, to keep only genes with more than 1 TPM on average across both groups, set the cutoff to 10 if there are 10 replicates in total.}

\item{gene_column}{A character string naming the column containing gene names.}

\item{batches}{A character vector identifying the batch structure of the experiment. The length must equal \code{length(group_1) + length(group_2)}. If \code{NULL}, there are no batches, or batches have already been corrected. Defaults to \code{NULL}.}

\item{batch_corr_method}{A character string naming the batch correction method. Can be one of \code{combat_np} (non-parametric) or \code{combat_p} (parametric). Defaults to \code{combat_np}. Ignored if \code{batches} is \code{NULL}.}

\item{bcorr_data_validation}{\code{NULL} if no batch corrected data is already available. Otherwise, a data frame of treatment-corrected data should be supplied (to speed up validation, if already available).}
}
\value{
An object of class \code{delboy}. Access this object using \code{delboy::hits}, \code{delboy::plot.delboy}, \code{delboy::get_performance_stats}, and \code{delboy::get_deseq2_results}.
}
\description{
Performs a differential-representation analysis using an elastic-net logistic regression approach for normalized count data that is split into two groups.
}
\references{
\itemize{
\item Kalinka, A. T. 2020. Improving the sensitivity of differential-expression analyses for under-powered RNA-seq experiments. bioRxiv.
\item Patro, R. et al. 2017. Salmon provides fast and bias-aware quantification of transcript expression. Nature Methods 14: 417-419.
}
}
\seealso{
\code{\link{hits}}, \code{\link{plot.delboy}}, \code{\link{get_performance_stats}}, \code{\link{get_deseq2_results}}
}
