% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_delboy.R
\name{run_delboy}
\alias{run_delboy}
\title{run_delboy}
\usage{
run_delboy(
  data,
  group_1,
  group_2,
  filter_cutoff,
  gene_column,
  batches = NULL,
  max.iter = 10,
  bcorr_data_validation = NULL,
  alpha = 0.5
)
}
\arguments{
\item{data}{A data frame containing normalized count data for two different groups and their replicates. Can be a path to a file. Ideally the counts will already have been normalized to Transcripts Per Million (TPM), using, for example, the bias-aware quantification methods employed by \code{salmon} (Patro et al. 2017).}

\item{group_1}{A character string naming the columns that belong to group 1.}

\item{group_2}{A character string naming the columns that belong to group 2.}

\item{filter_cutoff}{A numerical value indicating the cutoff below which (summed across all replicates) a gene will be removed from the data. For example, to keep only genes with more than 1 TPM on average across both groups, set the cutoff to 10 if there are 10 replicates in total.}

\item{gene_column}{A character string naming the column containing gene names.}

\item{batches}{A named character vector identifying the batch structure with names identifying sample columns in the data input. The length must equal \code{length(group_1) + length(group_2)}. If \code{NULL}, there are no batches, or batches have already been corrected. Defaults to \code{NULL}. Batch correction will be conducted using \code{sva::ComBat} using non-parametric priors.}

\item{max.iter}{An integer value indicating the maximum number of validation samples (default = 10). \code{NULL} indicates all sample combinations should be taken.}

\item{bcorr_data_validation}{\code{NULL} if no batch (signal) corrected data is already available for validation. Otherwise, a data frame of treatment-corrected data should be supplied (to speed up validation, if already available). Defaults to \code{NULL}. Batch correction will be conducted using \code{sva::ComBat} using non-parametric priors.}

\item{alpha}{The elastic-net regression penalty, between 0 and 1 (default = 0.5). If \code{NULL}, alpha is chosen automatically.}
}
\value{
An object of class \code{delboy}. Access this object using \code{hits}, \code{plot.delboy}, \code{get_performance_stats}, and \code{get_deseq2_results}.
}
\description{
Performs a differential-representation analysis using an elastic-net logistic regression approach for normalized count data that is split into two groups.
}
\references{
\itemize{
\item Kalinka, A. T. 2020. Improving the sensitivity of differential-expression analyses for under-powered RNA-seq experiments. bioRxiv \href{https://doi.org/10.1101/2020.10.15.340737}{10.1101/2020.10.15.340737}.
\item Patro, R. et al. 2017. Salmon provides fast and bias-aware quantification of transcript expression. Nature Methods 14: 417-419.
}
}
\seealso{
\code{\link{hits}}, \code{\link{plot.delboy}}, \code{\link{get_performance_stats}}, \code{\link{get_deseq2_results}}
}
\author{
Alex T. Kalinka, \email{alex.t.kalinka@gmail.com}
}
