% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_delboy_crispr.R
\name{run_delboy_crispr}
\alias{run_delboy_crispr}
\title{run_delboy_crispr}
\usage{
run_delboy_crispr(
  data,
  controls,
  treatments,
  grna_column,
  gene_column,
  target_fdr = 0.1,
  normalize_method = "relative",
  max.iter = 3,
  filter_prop = 0.05
)
}
\arguments{
\item{data}{A data frame containing count data for two different groups and their replicates.}

\item{controls}{A character string naming the control sample columns.}

\item{treatments}{A character string naming the treatment sample columns.}

\item{grna_column}{A character string naming the column containing gene names.}

\item{gene_column}{A character string naming the column containing gene names.}

\item{target_fdr}{Numeric value (0-1) giving the target FDR. Defaults to 0.1.}

\item{normalize_method}{A character string naming the read depth normalization method: \code{relative} (default) or \code{median_ratio}.}

\item{max.iter}{An integer value indicating the maximum number of validation sample combinations (default = 3). \code{NULL} indicates all sample combinations should be taken. \code{NULL} indicates that the validation step should be skipped.}

\item{filter_prop}{A numerical value (0-1) to filter the lowest abundance genes when estimating the non-null logFC distribution for performance estimation. Defaults to 0.05.}
}
\value{
An object of class \code{delboy_crispr}.
}
\description{
Calculates a threshold value for a given metric to achieve a specified FDR.
}
