% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_delboy_crispr.R
\name{run_delboy_crispr}
\alias{run_delboy_crispr}
\title{run_delboy_crispr}
\usage{
run_delboy_crispr(
  data,
  controls,
  treatments,
  filter_cutoff,
  grna_column,
  gene_column,
  target_fdr = 0.1,
  normalize_method = "relative",
  max.iter = 3
)
}
\arguments{
\item{data}{A data frame containing count data for two different groups and their replicates.}

\item{controls}{A character string naming the control sample columns.}

\item{treatments}{A character string naming the treatment sample columns.}

\item{filter_cutoff}{A numerical value indicating the cutoff below which (summed across all replicates) a gene will be removed from the data. For example, to keep only genes with more than 1 TPM on average across both groups, set the cutoff to 10 if there are 10 replicates in total.}

\item{grna_column}{A character string naming the column containing gene names.}

\item{gene_column}{A character string naming the column containing gene names.}

\item{target_fdr}{Numeric value (0-1) giving the target FDR. Defaults to 0.1.}

\item{normalize_method}{A character string naming the read depth normalization method: \code{relative} (default) or \code{median_ratio}.}

\item{max.iter}{An integer value indicating the maximum number of validation sample combinations (default = 3). \code{NULL} indicates all sample combinations should be taken. \code{NA} indicates that the validation step should be skipped.}
}
\value{
A data frame.
}
\description{
Calculates a threshold value for a given metric to achieve a specified FDR.
}
